This dataset contains information collected by the U.S Census Service concerning housing in the area of Boston Mass.

This dataset has 516 observations and 14 variables. It contains information collected by the U.S Census Service oncerning housing in the area of Boston Mass. It was obtained from the StatLib archive http://lib.stat.cmu.edu/datasets/boston, and has been used extensively throughout the literature to benchmark algorithms. Note that the details of this dataset were sourced from https://www.cs.toronto.edu/~delve/data/boston/bostonDetail.html. The data was originally published in see Harrison Jr and Rubinfeld (1978)

boston_housing

Format

A tibble with 516 rows and 14 variables:

crim

per capita crime rate by town

zn

proportion of residential land zoned for lots over 25,000 sq.ft.

indus

proportion of non-retail business acres per town.

chas

Charles River dummy variable (1 if tract bounds river; 0 otherwise)

nox

nitric oxides concentration (parts per 10 million)

rm

average number of rooms per dwelling

age

proportion of owner-occupied units built prior to 1940

dis

weighted distances to five Boston employment centres

rad

index of accessibility to radial highways

tax

full-value property-tax rate per $10,000

ptratio

pupil-teacher ratio by town

b

1000(Bk - 0.63)^2 where Bk is the proportion of blacks by town

lstat

lower status of the population

medv

Median value of owner-occupied homes in $1000's

Source

http://lib.stat.cmu.edu/datasets/boston

References

Harrison Jr D, Rubinfeld DL (1978). “Hedonic housing prices and the demand for clean air.” Journal of environmental economics and management, 5(1), 81--102.