Compute the sandwich estimator of standard errors for OLS

Compute the sandwich estimator of standard errors for ordinary least squares (OLS) regression, see White (1980) and White (1980) . For more details see also Buja et al. (2019) and Buja et al. (2019) .

comp_sand_var(mod_fit)

Arguments

mod_fit
A lm (OLS) object.

Value

A list containing the following elements: the type of estimator of of the variance (var_type); An abbreviated string representing the type of the estimator of the variance (var_type_abb); the summary statistics of mod_fit based on this estimator of the variance (e.g., standard errors and p-values) (var_summary); the assumptions under which the estimator of the variance is consistent (var_assumptions); the covariance matrix for the coefficients estimates (cov_mat).

Details

The function computes the sandwich estimator for the OLS regression passed in mod_fit and returns a tibble with summary information about the components of the model based on the sandwich standard errors.

References

Buja A, Brown L, Berk R, George E, Pitkin E, Traskin M, Zhang K, Zhao L (2019). “Models as approximations I: consequences illustrated with linear regression.” Statist. Sci., 34(4), 523--544. ISSN 0883-4237.

Buja A, Brown L, Kuchibhotla AK, Berk R, George E, Zhao L (2019). “Models as approximations II: A model-free theory of parametric regression.” Statist. Sci., 34(4), 545--565. ISSN 0883-4237.

White H (1980). “Using least squares to approximate unknown regression functions.” Internat. Econom. Rev., 21(1), 149--170. ISSN 0020-6598.

White H (1980). “A heteroskedasticity-consistent covariance matrix estimator and a direct test for heteroskedasticity.” Econometrica, 48(4), 817--838. ISSN 0012-9682.

Examples

if (FALSE) {
set.seed(674748)
n <- 1e5
X <- stats::rnorm(n, 0, 1)
y <- 2 + X * 1 + stats::rnorm(n, 0, 10)
lm_fit <- stats::lm(y ~ X)
sandwich_qr_std_err <- comp_sand_var(lm_fit)
}