Confidence interval of maars_lm, lm object

confint method for an object of class maars_lm, lm.

# S3 method for maars_lm
confint(
  object,
  parm = NULL,
  level = 0.95,
  sand = TRUE,
  boot_emp = FALSE,
  boot_sub = FALSE,
  boot_res = FALSE,
  boot_mul = FALSE,
  well_specified = FALSE,
  digits = 3,
  ...
)

Arguments

object
A fitted "maars_lm" object.

parm
(NULL) : a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered. Currently only allowed value is NULL.

level
(numeric) : TODO: Add later

sand
(logical) : TRUE if sandwich estimator output is required, FALSE to exclude this output from the request.

boot_emp
(logical) : TRUE if empirical bootstrap standard error output is required, FALSE to exclude this output from the request.

boot_sub
(logical) : TRUE if subsampling standard error output is required, FALSE to exclude this output from the request.

boot_res
(logical) : TRUE if residual bootstrap standard error output is required, FALSE to exclude this output from the request.

boot_mul
(logical) : TRUE if multiplier bootstrap standard error output is required, FALSE to exclude this output from the request

well_specified
(logical) : TRUE if lm standard errors. (well specified) output is required, FALSE to exclude this output from the request.

digits
(integer) : Rounding digits used in some of the function's output.

...
Additional arguments.