Generates list containing estimates for an individual variance type

get_mms_comp_var_ind generates the final output list for an individual variance type. This is a helper function to be used in the following functions: comp_lm_var, comp_sand_var, comp_boot_emp, comp_boot_mul, comp_boot_res.

get_mms_comp_var_ind(
  var_type_abb,
  summary_tbl = NULL,
  cov_mat = NULL,
  B = NULL,
  m = NULL,
  n = NULL,
  weights_type = NULL,
  boot_out = NULL
)

Arguments

var_type_abb
(character) : The abbreviation used to classify the type of variance. The choice is one of the following "lm", "sand", "emp", "sub", "res", "mul".

summary_tbl
(tibble) : The summary table of standard errors for the type of variance.

cov_mat
(matrix) : The covariance matrix for the type of variance.

B
(integer) : The number of bootstrap replications.

m
(integer) : The number of observations to be sampled with or without replacement from the dataset for each bootstrap repetition.

n
(integer) : The number of observations from the original modeling dataset.

weights_type
: The type of multiplier bootstrap weights to generate. Based on the weighttype option in the Stata boottest package, this can only take the following five prespecified values "rademacher", "mammen", "webb", "std_gaussian", "gamma". For more details see the documentation for comp_boot_mul_wgt. The default value is "rademacher".

boot_out
(tibble) : A tibble of the boostrap replication datasets used in comp_boot_mul, comp_boot_res. In all other variance types this should be set to NULL.

Value

(list) : A list containing essential computed variance output for the type of variance specified. The computed parameter types are the common output for the following functions: comp_lm_var, comp_sand_var, comp_boot_emp, comp_boot_mul, comp_boot_res.